requirements.txt
pandas
sqlalchemy
numpy
matplotlib
requests
scipy
pymysql
pyyaml